refactor : mvc 패턴 걷어내고 기능 위주로 폴더 구조 재정립
