+++ src/input/input-parser.js
@@ -1,0 +1,11 @@
+class InputParser {
+  static parseCarNameInput(input) {
+    return input.split(',');
+  }
+
+  static parseAttemptCountInput(input) {
+    return parseInt(input, 10);
+  }
+}
+
+export default InputParser;
